Stripe integration improvements: - Failed subscription charge webhook cancels subscription immediately instead of at period end - Stripe users email address is now also updated when admin changes a users email

+++ app/Billing/StripeBiller.php
@@ -1,8 +1,10 @@
 <?php namespace RentGorilla\Billing;
 
-use Stripe\Charge;
+use RentGorilla\User;
 use Stripe\Customer;
+use Stripe\Charge;
 use Stripe\Plan;
+use Log;
 
 class StripeBiller implements Biller
 {
@@ -25,15 +27,36 @@ class StripeBiller implements Biller
         return null;
     }
 
-    public function deleteAccount($customer_id)
+    public function deleteAccount(User $user)
     {
-        $cu = $this->getCustomerById($customer_id);
-        return $cu->delete();
+        if($user->readyForBilling()) {
+            try {
+                $cu = $this->getCustomerById($user->getStripeId());
+                $cu->delete();
+                Log::info('Deleted stripe user ' . $user->getStripeId());
+            } catch (\Exception $e) {
+                Log::error($e->getMessage());
+            }
+        }
     }
 
     public function getCustomerById($customer_id)
     {
         return Customer::retrieve($customer_id);
+    }
+
+    public function updateEmail(User $user)
+    {
+        if($user->readyForBilling()) {
+            try {
+                $cu = $this->getCustomerById($user->getStripeId());
+                $cu->email = $user->email;
+                $cu->save();
+                Log::info('Updated stripe user email ' . $user->email);
+            } catch (\Exception $e) {
+                Log::error($e->getMessage());
+            }
+        }
     }
 
 }
